Remove redundant files (#19)
* Use only a single file to store Np

We don't need to redundantly store Np as it is the same across all
the MPI processes.

* Remove `N_rank_%d_size_%d` files

These files are not used.

* Remove `Nelements_rank_%d_size_%d` files

These files are not used.

* Remove unused main_surrogate.py

main_surrogate.py is no longer needed and used so it can be removed

* Remove unused main_surrogate.py

main_surrogate.py is no longer needed and used so it can be removed

* Clean up writing of unused files from gnn plugin

Comment out writing of N and Nelements, and only write Np from rank 0, now matching what is being read by trainer.py

---------

Co-authored-by: Riccardo Balin <balin@aurora-uan-0011.hostmgmt.cm.aurora.alcf.anl.gov>
Co-authored-by: Riccardo Balin <riccardo.balin@gmail.com>